Add boosting terms to query
These is using some private API's that allow some custom boosting. We are iterating on getting this embedded into the service itself, but in the meanwhile, this should improve results substantially.  Please note that I do not know javascript and am guessing at style guidelines and syntax.....  Please feel free to play with the numbers some on your side as we work out our improved ranking.